requirements/base.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
alibabacloud-credentials==0.3.2
alibabacloud-dingtalk==2.0.11
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-dingtalk==1.0.2
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-tea==0.3.1
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pendulum==2.1.2
pip-review==1.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
requests==2.30.0
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
